requirements-dev.txt
# ══════════════════════════════════════════════════════════════
# requirements-dev.txt
# Dependencias para desarrollo, pruebas y linting.
# Incluye requirements.txt como base.
#
# Instalación:
#   pip install -r requirements-dev.txt
#
# IMPORTANTE: usar >= en lugar de == para que pip elija
# el wheel precompilado correcto según el sistema operativo.
# En Windows esto evita errores de compilación (cl.exe / gcc).
# ══════════════════════════════════════════════════════════════

# Incluye todas las dependencias de producción
-r requirements.txt

# ── Testing ───────────────────────────────────────────────────
pytest>=8.1.1
pytest-cov>=5.0.0

# Simulación de AWS en memoria (sin cuenta real, sin Docker)
moto[s3,glue,athena,lambda,sqs,kms]>=5.0.3

pytest-asyncio>=0.23.6

# ── Calidad de código ─────────────────────────────────────────
black>=24.3.0
flake8>=7.0.0
isort>=5.13.2

# Stubs de tipos para autocompletado en VS Code
boto3-stubs[s3,glue,athena,lambda,sqs]>=1.34.69

# ── Utilidades de desarrollo ──────────────────────────────────
pre-commit>=3.7.0
